feat: add new components on about page
