+++ src/components/AboutTabs/AboutAuthor.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Subtitle, Paragraph } from '../../UI/Titles';
+
+const AboutAuthor = () => (
+  <>
+    <Subtitle>Author</Subtitle>
+    <Paragraph>Will be added later</Paragraph>
+  </>
+);
+
+export default AboutAuthor;
