+++ src/components/AboutTabs/index.js
@@ -1,0 +1,7 @@
+import AboutAdminOrMentorFeatures from './AboutAdminOrMentorFeatures';
+import AboutAuthor from './AboutAuthor';
+import AboutDIMS from './AboutDIMS';
+import AboutMemberFeatures from './AboutMemberFeatures';
+import AboutRoles from './AboutRoles';
+
+export { AboutAdminOrMentorFeatures, AboutAuthor, AboutDIMS, AboutMemberFeatures, AboutRoles };
